添加自定义上拉刷新控件

+++ YSLIBS/Ys.Camera.Droid/Views/YsCameraX.cs
@@ -210,7 +210,7 @@ namespace Ys.Camera.Droid.Views
         /// 设置缩放
         /// </summary>
         /// <param name="zoomPercent"></param>
-        public void SetZoom(int zoomPercent)
+        public void SetZoom(float zoomPercent)
         {
             try
             {
